src/requirements.txt
alembic
aniso8601
attrs
blinker
click
Flask
Flask-Cors
flask-marshmallow
Flask-Migrate
flask-restx
Flask-SQLAlchemy
importlib_resources
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
Mako
MarkupSafe
marshmallow
marshmallow-sqlalchemy
numpy
packaging
pandas
psycopg2-binary
python-dateutil
python-dotenv
pytz
referencing
rpds-py
six
SQLAlchemy
typing_extensions
tzdata
Werkzeug